ReusableComponents/Utilities/Services/Database/requirements.txt
redis==5.2.1
pyodbc==5.1.0
python-dotenv==1.0.1
